EVA_apps/EKEELVideoAugmentation/src/flask-server/requirements.txt
Flask==2.1.0
youtube-transcript-api==0.4.1
Flask-HTTPAuth==4.7.0
PyJWT==2.1.0
bcrypt==3.2.0
flask-mongoengine==1.0.0
Flask-Mail==0.9.1
urllib3
#urllib3==1.24.2  
json5==0.8.5
rdflib==6.1.1
#rdflib==6.3.2
PyLD==2.0.3
pandas==1.3.0
conllu==4.4
werkzeug~=2.0.0
python-dotenv==0.21.1